Use relative paths
The Hardcoded path to /srv/speed.pypy.org/releases/xxxx means that
we inadvertingly update the code for the running instance while
we are deploying the new instance. The window is small but it
means we could get a mismatch.

This change makes the path relative to the settings.py so it always
uses the current revision, and it removes the need for our
forked application_python.