Fix forwarded hostname for pycon servers
It appears (https://github.com/caktus/pycon/issues/188) that requests
getting to the backend production servers are being identified to
Django as having been addressed to the staging server - Django is
picking up the hostname as staging-pycon.python.org instead of
us.pycon.org.

I'm guessing this is due to the wrong value for proxy_pass in the
nginx config. On production I see:

  proxy_pass http://staging-pycon.python.org;

I'm also guessing I can set the application name in PyCon's app.rb
to correct this.

Once this is corrected, Django's ALLOWED_HOSTS settings will need
to accept the new hostname, so I've updated that as well. (It's
okay for that to include all the possible hostnames.)